refactor: move tipos de filtros para um local comum entre DataTable e FilterBar
