+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/print/buildPrintProps.ts
@@ -8,9 +8,9 @@ import {
 } from "@/helpers/larguraColunasHipertensao";
 import type { FilterItem } from "@/services/lista-nominal/ListaNominal";
 import { PROFILE_ID, type ProfileIdValue } from "@/types/profile";
-import { coeqColumns } from "../CoeqDataTable/consts";
 import { filtersLabels } from "../CoeqFiltersBar/consts";
 import type { DiabetesAcfPrintItem } from "./diabetes/print.model";
+import { coeqColumns } from "../DataTable/modules/CoeqDataTable/consts";
 
 export const buildPrintProps = (
     list: AcfDashboardType,
